requirements.txt
Flask==3.1.1
flask_cors==6.0.1
google_api_python_client==2.172.0
moderngl==5.12.0
moviepy==1.0.3
numpy==2.3.0
openai==1.88.0
Pillow==11.2.1
protobuf==6.31.1
pygame==2.6.1
requests==2.32.4
python-dotenv==1.0.0